+++ CHANGELOG.md
@@ -29,6 +29,11 @@ With this method, the version is YEAR.MONTH.RELEASE. To increment this version, 
 - Added `prompt_template`.
 - CLI app: log_level is now a command line argument.
 - Template: Suport relative paths from the config file.
+- Interactive: Added optional ability to extract answer from response with `--extract`.
+- Interactive: Gracefully handle EOF.
+- Interactive & Template: Make context optional.
+- Interactive: Allow context to be added inline with `/context <context>`.
+- Interactive: A blank context, unsets the context.
 
 ### Changed